requirements.txt
gitpython
openai
rich
pytest
flask 